+++ automation/ptp_operator_cleanup.sh
@@ -8,11 +8,11 @@ source ./functions.sh
 parse_args $@
 
 echo "Removing PtpConfig ..."
-oc delete -f ${MANIFEST_DIR}/ptp-config.yaml || true
+[ -e ${MANIFEST_DIR}/ptp-config.yaml ] && oc delete -f ${MANIFEST_DIR}/ptp-config.yaml || true
 echo "Removing PtpConfig: done"
 
 echo "Restore chronyd.service ..."
-oc delete -f ${MANIFEST_DIR}/disable-chronyd.yaml
+[ -e ${MANIFEST_DIR}/disable-chronyd.yaml ] && oc delete -f ${MANIFEST_DIR}/disable-chronyd.yaml
 
 if [[ "${WAIT_MCP}" == "true" ]]; then
     wait_mcp
